day06/day06_test.go
package day06_test

import (
	"path"
	"testing"

	"github.com/chigley/advent2021"
	"github.com/chigley/advent2021/day06"
	"github.com/stretchr/testify/assert"
)

var tests = []struct {
	in    string
	part1 int
	part2 int
}{
	{"example", 5934, 26984457539},
	{"input", 371379, 1674303997472},
}

func TestDay06(t *testing.T) {
	for _, tt := range tests {
		tt := tt

		t.Run(tt.in, func(t *testing.T) {
			t.Parallel()

			in, err := advent2021.ReadIntLines(path.Join("testdata", tt.in))
			if err != nil {
				t.Fatal(err)
			}

			assert.Equal(t, tt.part1, day06.Part1(in[0]))
			assert.Equal(t, tt.part2, day06.Part2(in[0]))
		})
	}
}
